requirements.txt
python-dotenv
pandas 
numpy 
pymongo 
certifi
pymongo[srv]
ipykernel
scikit-learn
dill
pyaml
##-e.